requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-localflavor==4.0
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.1
pytest-django==4.8.0
python-decouple==3.8
python-stdnum==1.20
sqlparse==0.5.1
typing_extensions==4.12.2
